add stop_job function
