+++ backend/indexing/run_job.py
@@ -18,6 +18,7 @@ from indexing.run_management_tasks import (
 )
 from indexing.run_update_task import run_update_task
 from ianalyzer.celery_utils import warn_if_no_worker
+from indexing.stop_job import mark_tasks_stopped
 
 
 logger = logging.getLogger('indexing')
@@ -58,15 +59,6 @@ def run_task(self, task: IndexTask) -> None:
 
     if isinstance(task, CreateIndexTask):
         task.client().cluster.health(wait_for_status='yellow')
-
-
-def mark_tasks_stopped(job: IndexJob):
-    '''
-    Mark open tasks as aborted and queued tasks as cancelled.
-    '''
-    for task_set in job.task_query_sets():
-        task_set.filter(status=TaskStatus.QUEUED).update(status=TaskStatus.CANCELLED)
-        task_set.filter(status=TaskStatus.WORKING).update(status=TaskStatus.ABORTED)
 
 
 @celery.shared_task()
